Dev Eco (#11)
* update compose to create a local network
update code to be consistent with env variables and names used in gleaner/nabu/glcon


* set headless to a container version that works


* Test to see if we can add report


* upload release


* add reports,
catch container issues #9
upload release #5



* fix some of the env variable mapping



* Debug flag

* Graph Endpoint

* fix missing graph


* rename containers



* Notes and archive


* add env for volumes
fix an issue with logging

+++ dagster/implnets/templates/v1/implnet_ops_SOURCEVAL.py
@@ -10,7 +10,7 @@ from ec.gleanerio.gleaner import getGleaner, getSitemapSourcesFromGleaner
 from minio import Minio
 from minio.error import S3Error
 from datetime import datetime
-from ec.reporting.report import missingReport, generateGraphReportsRepo
+from ec.reporting.report import missingReport, generateGraphReportsRepo, reportTypes
 from ec.datastore import s3
 from ec.graph.manageGraph import ManageBlazegraph as mg
 
@@ -22,6 +22,7 @@ from ec.reporting.report import missingReport
 from ec.datastore import s3
 
 DEBUG=os.environ.get('DEBUG')
+GLEANER_CONFIG_VOLUME=os.environ.get('GLEANER_CONFIG_VOLUME')
 # Vars and Envs
 GLEANER_HEADLESS_NETWORK=os.environ.get('GLEANER_HEADLESS_NETWORK')
 # env items
@@ -275,9 +276,10 @@ def gleanerio(mode, source):
 
         data["Env"] = enva
         data["HostConfig"] = {
-            "NetworkMode": GLEANER_HEADLESS_NETWORK
-            }
-        # data["Binds"] = [
+            "NetworkMode": GLEANER_HEADLESS_NETWORK,
+            "Binds":  [f"{GLEANER_CONFIG_VOLUME}:/configs"]
+        }
+        # data["Volumes"] = [
         #     "dagster-project:/configs"
         # ]
         # we would like this to be "dagster-${PROJECT:-eco}" but that is a bit tricky
@@ -316,7 +318,10 @@ def gleanerio(mode, source):
                 print("failed to create container:  unknown reason: ", err)
                 get_dagster_logger().info(f"Create Failed: unknown reason {str(err)}")
             raise err
-
+        except Exception as err:
+            print("failed to create container:  unknown reason: ", err)
+            get_dagster_logger().info(f"Create Failed: unknown reason {str(err)}")
+            raise err
         # print(cid)
 
         ## ------------  Archive to load, which is how to send in the config (from where?)
@@ -399,17 +404,40 @@ def gleanerio(mode, source):
 
         get_dagster_logger().info(f"container Logs to s3: {str(r.status)}")
 
+## get log files
+
         ## ------------  Remove   expect 204
+        url = URL + 'containers/' + cid + '/archive'
+        params = {
+            'path': f"{WorkingDir}/logs"
+        }
+        query_string = urllib.parse.urlencode(params)
+        url = url + "?" + query_string
+
+        # print(url)
+        req = request.Request(url,  method="GET")
+        req.add_header('X-API-Key', APIKEY)
+        req.add_header('content-type', 'application/x-compressed')
+        req.add_header('accept', 'application/json')
+        r = request.urlopen(req)
+
+        log.info(f"{r.status} ")
+        get_dagster_logger().info(f"Container Archive Retrieved: {str(r.status)}")
+       # s3loader(r.read().decode('latin-1'), NAME)
+        s3loader(r.read(), f"{source}_runlogs")
     finally:
         if (not DEBUG) :
-            url = URL + 'containers/' + cid
-            req = request.Request(url, method="DELETE")
-            req.add_header('X-API-Key', APIKEY)
-            # req.add_header('content-type', 'application/json')
-            req.add_header('accept', 'application/json')
-            r = request.urlopen(req)
-            print(r.status)
-            get_dagster_logger().info(f"Container Remove: {str(r.status)}")
+            if (cid):
+                url = URL + 'containers/' + cid
+                req = request.Request(url, method="DELETE")
+                req.add_header('X-API-Key', APIKEY)
+                # req.add_header('content-type', 'application/json')
+                req.add_header('accept', 'application/json')
+                r = request.urlopen(req)
+                print(r.status)
+                get_dagster_logger().info(f"Container Remove: {str(r.status)}")
+            else:
+                get_dagster_logger().info(f"Container Not created, so not removed.")
         else:
             get_dagster_logger().info(f"Container NOT Remove: DEBUG ENABLED")
 
@@ -499,7 +527,7 @@ def SOURCEVAL_graph_reports(context, msg: str):
 
     milled = False
     summon = True
-    returned_value = generateGraphReportsRepo(source_name,  graphendpoint)
+    returned_value = generateGraphReportsRepo(source_name,  graphendpoint, reportList=reportTypes["repo_detailed"])
     r = str('returned value:{}'.format(returned_value))
     #report = json.dumps(returned_value, indent=2) # value already json.dumps
     report = returned_value
